+++ pkg/scheduler/api/common_info/errors.go
@@ -19,6 +19,7 @@ const (
 	ResourcesWereNotFoundMsg = "no nodes with enough resources were found"
 	DefaultPodgroupError     = "Unable to schedule podgroup"
 	DefaultPodError          = "Unable to schedule pod"
+	OverheadMessage          = "Not enough resources due to pod overhead resources"
 )
 
 type FitError struct {
@@ -60,7 +61,7 @@ func NewFitErrorByReasons(name, namespace, nodeName string, err error, reasons .
 func NewFitErrorInsufficientResource(
 	name, namespace, nodeName string,
 	resourceRequested *resource_info.ResourceRequirements, usedResource, capacityResource *resource_info.Resource,
-	capacityGpuMemory int64, gangSchedulingJob bool,
+	capacityGpuMemory int64, gangSchedulingJob bool, messageSuffix string,
 ) *FitError {
 	availableResource := capacityResource.Clone()
 	availableResource.Sub(usedResource)
@@ -143,6 +144,15 @@ func NewFitErrorInsufficientResource(
 				gangSchedulingJob))
 			shortMessages = append(shortMessages, fmt.Sprintf("node(s) didn't have enough resources: %s",
 				requestedResourceName))
+		}
+	}
+
+	if len(messageSuffix) > 0 {
+		for i, msg := range shortMessages {
+			shortMessages[i] = fmt.Sprintf("%s. %s", msg, messageSuffix)
+		}
+		for i, msg := range detailedMessages {
+			detailedMessages[i] = fmt.Sprintf("%s. %s", msg, messageSuffix)
 		}
 	}